+++ trunk/src/common/tortls.h
@@ -27,10 +27,10 @@ int tor_tls_context_new(crypto_pk_env_t *rsa, int isServer,
 tor_tls *tor_tls_new(int sock, int is_server, int use_no_cert);
 void tor_tls_free(tor_tls *tls);
 int tor_tls_peer_has_cert(tor_tls *tls);
-int tor_tls_get_peer_cert_nickname(tor_tls *tls, char *buf, int buflen);
+int tor_tls_get_peer_cert_nickname(tor_tls *tls, char *buf, size_t buflen);
 int tor_tls_verify(tor_tls *tls, crypto_pk_env_t **identity);
-int tor_tls_read(tor_tls *tls, char *cp, int len);
-int tor_tls_write(tor_tls *tls, char *cp, int n);
+int tor_tls_read(tor_tls *tls, char *cp, size_t len);
+int tor_tls_write(tor_tls *tls, char *cp, size_t n);
 int tor_tls_handshake(tor_tls *tls);
 int tor_tls_shutdown(tor_tls *tls);
 int tor_tls_get_pending_bytes(tor_tls *tls);
